# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API สำหรับ blog — อ่านได้ทุกคน แก้ได้เฉพาะเจ้าของ",
        "title": "Blog User API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/blogs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blog"
                ],
                "summary": "list blog",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "หน้า เริ่มที่ 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "ต่อหน้า สูงสุด 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ค้นจาก title/content",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "created_at",
                            "title"
                        ],
                        "type": "string",
                        "default": "created_at",
                        "description": "คอลัมน์ที่ใช้เรียง",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "ทิศทางการเรียง",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "กรองตามผู้เขียน (UUID)",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_dto_response.BlogResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "sort/order/user_id นอก whitelist",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blog"
                ],
                "summary": "สร้าง blog",
                "parameters": [
                    {
                        "description": "title กับ content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_dto_request.CreateBlogRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_dto_response.BlogResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    }
                }
            }
        },
        "/api/v1/blogs/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blog"
                ],
                "summary": "ดู blog ตาม id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "blog id (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_dto_response.BlogResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "id ไม่ใช่ UUID",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blog"
                ],
                "summary": "แก้ blog",
                "parameters": [
                    {
                        "type": "string",
                        "description": "blog id (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ต้องส่ง title กับ content ครบ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_dto_request.UpdateBlogRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_dto_response.BlogResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    },
                    "403": {
                        "description": "ไม่ใช่เจ้าของ",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blog"
                ],
                "summary": "ลบ blog",
                "parameters": [
                    {
                        "type": "string",
                        "description": "blog id (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    },
                    "403": {
                        "description": "ไม่ใช่เจ้าของ",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_thitiphongD_blog-user-api_internal_dto_request.CreateBlogRequest": {
            "type": "object",
            "required": [
                "content",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "minLength": 1
                },
                "title": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                }
            }
        },
...
        "github_com_thitiphongD_blog-user-api_internal_dto_request.UpdateBlogRequest": {
            "type": "object",
            "required": [
                "content",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "minLength": 1
                },
                "title": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                }
            }
        },
...
        "github_com_thitiphongD_blog-user-api_internal_dto_response.AuthorResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_thitiphongD_blog-user-api_internal_dto_response.BlogResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_dto_response.AuthorResponse"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_thitiphongD_blog-user-api_internal_response.Body": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Pagination"
                },
                "request_id": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "github_com_thitiphongD_blog-user-api_internal_response.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "ใส่ \"Bearer \" นำหน้า token ที่ได้จาก /api/v1/auth/login",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}